+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.45"
+__version__ = "0.12.46"
 
 
 def safe_commit() -> None: